+++ lib/neo4j/shared/typecasting/boolean.rb
@@ -1,7 +1,0 @@
-module Neo4j::Shared
-  module Typecasting
-    # Boolean provides a constant to use in typecaster resolution
-    class Boolean
-    end
-  end
-end