+++ lib/neo4j/shared/typecasting/unknown_typecaster_error.rb
@@ -1,8 +1,0 @@
-module Neo4j::Shared
-  module Typecasting
-    # This exception is raised if attempting to cast to an unknown type when
-    # using {Typecasting}
-    class UnknownTypecasterError < TypeError
-    end
-  end
-end